main/Agents/requirements.txt
google-adk==0.5.0
google-generativeai
langchain_google_genai==1.0.10
fastapi==0.115.9
protobuf==6.31.0
pydantic==2.8.2
python-dotenv==1.1.0
uvicorn==0.30.6


